refactor(test_functions): migrate `SpeedReducerShaft` to YAML spec; rename to `RSSpeedReducerShaft`
Migrate the speed reducer shaft function from the legacy ABC-based
class hierarchy to a YAML-driven spec. The function is fixed-dimension
(5D) with no parameters. It is also renamed to follow the `RS*`
convention for R-S reliability problems, consistent with
`RSCircularBar` and `RSQuadratic`.

The Gumbel (max) moment-to-parameter conversion is expressed
via the extended `$()` resolver using the newly whitelisted helpers
`gumbel_max_mu` and `gumbel_max_beta`, keeping the published
mean and standard deviation values visible in the spec.

- Add `rs_speed_reducer_shaft.yaml`
- Strip `speed_reducer_shaft.py` to `evaluate()` only
  and rename to `rs_speed_reducer_shaft.py`
- Rename docs page to `rs-speed-reducer-shaft.md`
  and update the MyST label and TOC entry
- Apply standard doc updates: revised opening paragraph,
  expanded prose footnote for SI unit conversions,
  raw string prefix on matplotlib labels,
  capitalized footnote sentence
- Add Li et al. (2018) MCS reference to
  the failure probability table
- Fix variable order in the formula's vector definition
  to match the marginals ordering

Partially resolves: #515
Ref: #502
